Fixed the bugs that I had, relocated ports and removed useless files such as duplicate entities, and the SLN file for suggestionservice.
