+++ git-together.sln
@@ -9,9 +9,13 @@ Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "ProfileService", "projects\
 EndProject
 Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "Domain", "shared\Domain\Domain.csproj", "{C9060E6A-6B7C-4638-8C97-7A8F888D7E13}"
 EndProject
-Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "Suggestion", "projects\Suggestion\Suggestion\Suggestion.csproj", "{35AB3134-505D-466E-9F8D-042D055DFF09}"
+Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "SuggestionService", "projects\Suggestion\Suggestion\SuggestionService.csproj", "{35AB3134-505D-466E-9F8D-042D055DFF09}"
 EndProject
 Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "Messages", "shared\Messages\Messages.csproj", "{F1493A7F-117E-4CD3-BF99-3B18BD7FDE83}"
+EndProject
+Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "Blacklist", "projects\Blacklist\Blacklist\Blacklist.csproj", "{AA9A4981-EE69-4DA3-AC10-587E0E6EC4FB}"
+EndProject
+Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "Messaging", "projects\Blacklist\Messaging\Messaging.csproj", "{C161AE6F-4E5E-495B-98BA-89AB7B6669A8}"
 EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
@@ -39,6 +43,14 @@ Global
 		{F1493A7F-117E-4CD3-BF99-3B18BD7FDE83}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{F1493A7F-117E-4CD3-BF99-3B18BD7FDE83}.Release|Any CPU.ActiveCfg = Release|Any CPU
 		{F1493A7F-117E-4CD3-BF99-3B18BD7FDE83}.Release|Any CPU.Build.0 = Release|Any CPU
+		{AA9A4981-EE69-4DA3-AC10-587E0E6EC4FB}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{AA9A4981-EE69-4DA3-AC10-587E0E6EC4FB}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{AA9A4981-EE69-4DA3-AC10-587E0E6EC4FB}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{AA9A4981-EE69-4DA3-AC10-587E0E6EC4FB}.Release|Any CPU.Build.0 = Release|Any CPU
+		{C161AE6F-4E5E-495B-98BA-89AB7B6669A8}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{C161AE6F-4E5E-495B-98BA-89AB7B6669A8}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{C161AE6F-4E5E-495B-98BA-89AB7B6669A8}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{C161AE6F-4E5E-495B-98BA-89AB7B6669A8}.Release|Any CPU.Build.0 = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(SolutionProperties) = preSolution
 		HideSolutionNode = FALSE
