--- projects/Suggestion/Suggestion.sln
@@ -1,25 +1,0 @@
-﻿
-Microsoft Visual Studio Solution File, Format Version 12.00
-# Visual Studio Version 17
-VisualStudioVersion = 17.9.34622.214
-MinimumVisualStudioVersion = 10.0.40219.1
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Suggestion", "Suggestion\Suggestion.csproj", "{AC0414F8-F071-4AD7-8028-ECA0ABCB49C8}"
-EndProject
-Global
-	GlobalSection(SolutionConfigurationPlatforms) = preSolution
-		Debug|Any CPU = Debug|Any CPU
-		Release|Any CPU = Release|Any CPU
-	EndGlobalSection
-	GlobalSection(ProjectConfigurationPlatforms) = postSolution
-		{AC0414F8-F071-4AD7-8028-ECA0ABCB49C8}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{AC0414F8-F071-4AD7-8028-ECA0ABCB49C8}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{AC0414F8-F071-4AD7-8028-ECA0ABCB49C8}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{AC0414F8-F071-4AD7-8028-ECA0ABCB49C8}.Release|Any CPU.Build.0 = Release|Any CPU
-	EndGlobalSection
-	GlobalSection(SolutionProperties) = preSolution
-		HideSolutionNode = FALSE
-	EndGlobalSection
-	GlobalSection(ExtensibilityGlobals) = postSolution
-		SolutionGuid = {D3C53DC2-B0D5-4142-857A-2442C387968C}
-	EndGlobalSection
-EndGlobal
