--- projects/Suggestion/Suggestion/Core/Entities/Suggestion.cs
@@ -1,12 +1,0 @@
-﻿using Suggestion.Core.Entities;
-
-namespace Suggestion.Core.Entiies
-{
-    public class Suggestion
-    {
-        public int Id { get; set; }
-        public List<RankProfile> ProfileRankings { get; set; }
-
-        public int userID { get; set; }
-    }
-}
