+++ projects/Suggestion/Suggestion/Core/Repository/IRepository.cs
@@ -1,13 +1,15 @@
-﻿using Suggestion.Core.Entities;
+﻿
 
-namespace Suggestion.Core.Repository
+using Domain;
+
+namespace SuggestionService.Core.Repository
 {
     public interface IRepository
     {
 
-        public Entiies.Suggestion CreateSuggestion(Entiies.Suggestion);
+        public Suggestion CreateSuggestion(Suggestion suggestion);
 
-        public List<Entiies.Suggestion> GetSuggestions();
+        public List<Suggestion> GetSuggestions();
 
     }
 }
