+++ projects/Suggestion/Suggestion/SuggestionService.csproj
@@ -17,4 +17,8 @@
     <PackageReference Include="Swashbuckle.AspNetCore" Version="6.4.0" />
   </ItemGroup>
 
+  <ItemGroup>
+    <ProjectReference Include="..\..\..\shared\Domain\Domain.csproj" />
+  </ItemGroup>
+
 </Project>
