+++ shared/Domain/RankProfile.cs
@@ -1,6 +1,10 @@
-﻿using static System.Runtime.InteropServices.JavaScript.JSType;
+﻿using System;
+using System.Collections.Generic;
+using System.Linq;
+using System.Text;
+using System.Threading.Tasks;
 
-namespace Suggestion.Core.Entities
+namespace Domain
 {
     public class RankProfile
     {
