+++ shared/Domain/Suggestion.cs
@@ -1,0 +1,11 @@
+﻿
+namespace Domain;
+
+public class Suggestion
+{
+    public int Id { get; set; }
+    public List<RankProfile> ProfileRankings { get; set; }
+
+    public int userID { get; set; }
+}
+
